requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
meraki==1.37.2
multidict==6.0.4
packaging==23.1
Pillow==10.0.1
pip-review==1.3.0
Pygments==2.16.1
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
rich==13.5.2
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
wifi-qrcode-generator==1.0.1
yarl==1.9.2
